1.0.5
Added flow action for saving motion alert videos to an SMB share using the Video-ID from motion alert trigger.
Added flow trigger for liveview videos (only possible with cloud subscription). This trigger allows the usage of flow action to store the video on a SMB share. This trigger is not avtivating the motion alert.